services/buy-executor/requirements.txt
# Buy Executor Service
# Python 3.11
# Based on my-little-jennie/agent_job/requirements.txt

# Flask & Gunicorn
flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1

# Oracle DB
mysql-connector-python>=8.0.0
SQLAlchemy==2.0.31

# Data Processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# LLM
google-generativeai==0.8.5

# HTTP Requests & WebSocket
requests==2.31.0
websocket-client==1.8.0

# Utils
pytz
filelock==3.14.0
redis
